add test filter_by_state()
