+++ tests/conftest.py
@@ -16,12 +16,3 @@ date_test = [
             ("2024-03-11T02:26:18.671407", "11.03.2024"),
             ("2025-02-14T18:26:25.39273", "14.02.2025")
            ]
-
-@pytest.fixture()
-def processing_test():
-    return [
-            {'id': 41428829, 'state': 'EXECUTED', 'date': '2019-07-03T18:35:29.512364'},
-            {'id': 939719570, 'state': 'EXECUTED', 'date': '2018-06-30T02:08:58.425572'},
-            {'id': 594226727, 'state': 'CANCELED', 'date': '2018-09-12T21:27:25.241689'},
-            {'id': 615064591, 'state': 'CANCELED', 'date': '2018-10-14T08:21:33.419441'}
-           ]
